+++ modules/apps/collaboration/blogs/blogs-api/src/main/java/com/liferay/blogs/service/BlogsEntryLocalService.java
@@ -421,7 +421,7 @@ public interface BlogsEntryLocalService extends BaseLocalService,
 	/**
 	* @deprecated As of 1.1.0, with no direct replacement
 	*/
-	@java.lang.Deprecated
+	@Deprecated
 	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
 	public List<BlogsEntry> getNoAssetEntries();
